abstact chunking classes
