+++ redbox/redbox/loader/chunking/chunkers/tabular.py
@@ -1,0 +1,34 @@
+import logging
+from typing import Iterator
+from datetime import UTC, datetime
+
+from langchain_core.documents import Document
+
+from redbox.models.chain import GeneratedMetadata
+from redbox.loader.chunking.base import BaseChunker
+
+logger = logging.getLogger(__name__)
+
+
+class TabularDocumentChunker(BaseChunker):
+    def tabular_chunks(
+        self,
+        s3_key: str,
+        tabular_elements: list[dict[str, str]],
+        generated_metadata: GeneratedMetadata,
+    ) -> Iterator[Document]:
+        created_datetime = datetime.now(UTC)
+
+        for idx, el in enumerate(tabular_elements or []):
+            metadata = self._build_metadata(
+                index=idx,
+                s3_key=s3_key,
+                page_number=1,
+                created_datetime=created_datetime,
+                text=el["text"],
+                generated_metadata=generated_metadata,
+            )
+
+            metadata = {**metadata, **el.get("metadata", {})}
+
+            yield Document(page_content=el["text"], metadata=metadata)
